% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dect_fac.R
\name{dect_fac}
\alias{dect_fac}
\title{Detect factors whether the factor is truly a character.}
\usage{
dect_fac(x)
}
\arguments{
\item{x}{factor}
}
\value{
character
}
\description{
Detect factors whether the factor is truly a character.
}
\examples{
a<-factor(c("1", "2", "3"))
b<-factor(c("1","2","2"))

# It is a chracter not factor
dect_fac(a)

# It is a factor
dect_fac(b)

}
