% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_des.R
\name{reorder_desc}
\alias{reorder_desc}
\title{Reorder a factor using desc from dplyr}
\usage{
reorder_desc(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor; gives reordered levels in decending alphabetical order
}
\description{
This function allows you to reorder the levels of a factor in decreasing order. R defines decreasing
order as decreasing alphabetical order when the levels of factors are strings.
}
\examples{

x <- factor(c("a", "b", "c", "d"))
levels(x)
levels(reorder_desc(x))
}
