% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_level.R
\name{set_level}
\alias{set_level}
\title{Set levels of a factor to the order appeared in data}
\usage{
set_level(x)
}
\arguments{
\item{x}{a vector or a factor}
}
\value{
a factor with levels in the order appeared in x.
This function will only return unique x levels,
}
\description{
Set levels of a factor to the order appeared in data
}
\examples{
set_level(c("pear","apple","banana","banana","apple"))
}
