% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\alias{cube}
\alias{reciprocal}
\title{Apply a useful power}
\usage{
square(x, plot_it = FALSE)

cube(x, plot_it = FALSE)

reciprocal(x, plot_it = FALSE)
}
\arguments{
\item{x}{The vector to be squared or cubed.}

\item{plot_it}{Display a plot of \code{x} vs the output? Use logical.
\code{FALSE} by default.}
}
\value{
A vector that is the:
\itemize{
     \item square (for \code{square})
     \item cube (for \code{cube})
     \item reciprocal/inverse (for \code{reciprocal})
}
 of \code{x}.
}
\description{
That's it -- these functions just square, cube, or take the
reciprocal/inverse of a vector.
}
\details{
If you really want to see more, check out the internal \code{\link{pow}}
function that these functions depend on.
}
\examples{
cube(tenvec)
square(-5)
reciprocal(2)
}
